# PetClinicAPI/PetClinicAPI/Models/Requests/UpdatePetRequest.cs
namespace PetClinicAPI.Models.Requests
{
    public class UpdatePetRequest
    {
        public int PetId { get; set; }
        public int ClientId { get; set; }
        public string Name { get; set; }
        public DateTime Birthday { get; set; }
    }
}